pyproject.toml
[tool.poetry]
authors = ["gladson <gladsonbrito@gmail.com>"]
description = ""
name = "base_app"
packages = [{include = "core", from = "src"}, {include = "apps", from = "src"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = {extras = ["tz"], version = "^1.13.1"}
cryptography = "^42.0.5"
fastapi = {extras = ["all"], version = "^0.110.0"}
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.6.3"}
pydantic-settings = "^2.2.1"
python = "^3.12"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
uvicorn = "^0.27.1"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
httpx = "^0.27.0"
isort = "^5.13.2"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
ruff = "^0.3.0"
taskipy = "^1.12.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
exclude = [
  "migrations",
]
line-length = 79

[tool.isort]
extend_skip = ["migrations"]
line_length = 79
profile = "black"

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-skip = ["migrations"]

[tool.taskipy.tasks]
format = "blue . && isort ."
lint = "ruff check . && blue --check . --diff && isort --check . --diff"
post_test = "coverage html"
pre_test = "task lint && task format"
run = "uvicorn src.server:app --host 0.0.0.0 --reload"
test = "pytest -s -x --cov=src -vv"
